Deprecate OptionSetInfo, prepare way for new ChoiceInfo function (#2235)
OptionSetInfo only returns one value, the logical name of an option set
choice. We'd like to redesign this function to accommodate:

1. The logical name is a text wrapped number for number and Boolean
backed option sets coming from Dataverse. We recently added the ability
to get at these values with the Value and Boolean functions which is a
better way, strongly typed.
https://github.com/microsoft/Power-Fx/pull/2230
1. There are other pieces of information that are worth knowing, for
example the color of an option set choice, that can't be accommodated
with this design.
1. Our preferred design is to return an extensible record of
information. We could add an additional second argument with an enum to
determine what information to return, with a default value of logical
name, but this isn't our first choice.
1. We don't use the term OptionSet any longer in Dataverse, it has been
replaced with Choice, which is more consistent with our existing Choices
function.

As some hosts may still be using this function, it will remain available
as a function that can be enabled by calling config.EnableOptionSetInfo.
Value/Boolean can be used immediately as a workaround, wrapped with Text
if needed. We plan to add a new ChoiceInfo function with a separate PR.
