+++ src/libraries/Microsoft.PowerFx.Interpreter/Environment/PowerFxConfigExtensions.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using Microsoft.PowerFx.Core.Functions;
+using Microsoft.PowerFx.Core.Texl.Builtins;
 using Microsoft.PowerFx.Functions;
 using Microsoft.PowerFx.Interpreter;
 
@@ -59,6 +60,12 @@ namespace Microsoft.PowerFx
                 config.SymbolTable.AddFunction(func.Key);
                 config.AdditionalFunctions.Add(func.Key, func.Value);
             }
+        }
+
+        [Obsolete("OptionSetInfo function is deprecated. Use the Value function on an option set backed by a number and the Boolean function on an option set backed by a Boolean instead. A new ChoiceInfo function is in the works for access to logical names.")]
+        public static void EnableOptionSetInfo(this PowerFxConfig powerFxConfig)
+        {
+            powerFxConfig.AddFunction(new OptionSetInfoFunction());
         }
     }
 }
