+++ src/libraries/Microsoft.PowerFx.Interpreter/Functions/Library.cs
@@ -1236,17 +1236,6 @@ namespace Microsoft.PowerFx.Functions
                 NoErrorHandling(Now)
             },
             {
-                BuiltinFunctionsCore.OptionsSetInfo,
-                StandardErrorHandling<OptionSetValue>(
-                    BuiltinFunctionsCore.OptionsSetInfo.Name,
-                    expandArguments: NoArgExpansion,
-                    replaceBlankValues: DoNotReplaceBlank,
-                    checkRuntimeTypes: ExactValueTypeOrBlank<OptionSetValue>,
-                    checkRuntimeValues: DeferRuntimeValueChecking,
-                    returnBehavior: ReturnBehavior.ReturnEmptyStringIfAnyArgIsBlank,
-                    targetFunction: OptionSetValueToLogicalName)
-            },
-            {
                 BuiltinFunctionsCore.Or,
                 Or
             },
